requirements.txt
renpy
pygame-ce